+++ src/clindex/api.clj
@@ -164,7 +164,7 @@
                           (dissoc ::ns-track/unload ::ns-track/load)) ;; we can discard this first time since we are indexing everything
               tx-data (-> (indexer/all-facts {:projects all-projs
                                               :namespaces all-ns})
-                          #_utils/check-facts)]
+                          utils/check-facts)]
           (reset! effective-schema (merge schema extra-schema))
           (swap! db-conns assoc p (d/create-conn @effective-schema))
           (swap! all-projects-by-platform assoc p all-projs)